requirements.txt
langchain>=0.0.352
langchain-community>=0.0.1
fastapi>=0.104.1
pydantic>=2.5.2
sqlite3-api>=0.1.0
ollama>=0.1.3
requests>=2.31.0
beautifulsoup4>=4.12.2
python-dotenv>=1.0.0
faiss-cpu>=1.7.4
numpy>=1.26.2
pandas>=2.1.3
scikit-learn>=1.3.2
uvicorn>=0.24.0